pledge/requirements.txt
netconf
pyopenssl
cryptography==2.4.2
pycryptodome

#used for GRASP
cbor
netifaces

requests
